+++ src/main/java/com/mulmeong/reward/badge/application/BadgeService.java
@@ -1,8 +1,8 @@
 package com.mulmeong.reward.badge.application;
 
-import com.mulmeong.reward.badge.dto.BadgeDto;
 import com.mulmeong.reward.badge.dto.in.BadgeCreateDto;
 import com.mulmeong.reward.badge.dto.in.BadgeUpdateDto;
+import com.mulmeong.reward.badge.dto.out.BadgeDto;
 
 import java.util.List;
 
@@ -10,11 +10,12 @@ public interface BadgeService {
 
     void createBadge(BadgeCreateDto badgeCreateDto);
 
+    void updateBadge(BadgeUpdateDto badgeUpdateDto);
+
+    void deleteBadge(Long badgeId);
+
     BadgeDto getBadge(Long badgeId);
 
     List<BadgeDto> getAllBadges();
 
-    void updateBadge(BadgeUpdateDto badgeUpdateDto);
-
-    void deleteBadge(Long badgeId);
 }
